pyproject.toml
[tool.poetry]
name = "cl-adaptation"
version = "0.1.0"
description = ""
authors = ["Piotr <piotr.hondra@allegro.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# torch = "1.11"
# torchvision = "*"
# matplotlib = "*"
# numpy = "^1"
# seaborn = "*"
# tensorboard = "*"
# wandb = "*"
# python-dotenv = "*"
# imagecorruptions = "*"
# hydra-core = "^1.3.2"
# torch = "^2.5.1"
torch = "2.0.1"
torchvision = "0.15.2"
wand = "^0.6.13"
python-dotenv = "^1.0.1"
imagecorruptions = "^1.1.2"
hydra-core = "^1.3.2"
matplotlib = "^3.10.0"
hydra-submitit-launcher = "^1.2.0"
wandb = "^0.19.3"
seaborn = "^0.13.2"
ipykernel = "^6.29.5"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
